Corrige migrations de vendas (pagamentos, endereços, permissões)
Migration 078 (vendas_pagamentos):
- Corrige formas_pagamento → forma_de_pagamento
- Corrige plano_contas → plano_de_contas

Migration 079 (vendas_enderecos):
- Corrige tipo cidade_id: BIGINT → INT UNSIGNED
- Compatibiliza com tabela cidades que usa INT UNSIGNED

Migration 081 (permissões):
- Corrige estrutura do INSERT para usar colunas corretas
- Altera de (modulo, acao) para (nome, codigo, descricao, modulo)
- Usa padrão 'venda.acao' para código das permissões
- Adiciona criado_em com NOW()

+++ database/migrations/079_criar_tabela_vendas_enderecos.sql
@@ -17,8 +17,8 @@ CREATE TABLE IF NOT EXISTS vendas_enderecos (
     venda_id BIGINT UNSIGNED NOT NULL,
     venda_external_id VARCHAR(50) DEFAULT NULL,
 
-    -- Cidade
-    cidade_id BIGINT UNSIGNED DEFAULT NULL,
+    -- Cidade - INT UNSIGNED pois cidades usa INT
+    cidade_id INT UNSIGNED DEFAULT NULL,
     cidade_external_id VARCHAR(50) DEFAULT NULL,
 
     -- ========================================
